graphql-django/requirements.txt
Django==1.11
pytz==2017.2
djangorestframework==3.6
psycopg2
markdown
graphene_django
django-cors-headers
raven